+++ src/main/scala/edu/ie3/powerFactory2psdm/converter/GridConverter.scala
@@ -6,11 +6,17 @@
 
 package edu.ie3.powerFactory2psdm.converter
 
+import edu.ie3.datamodel.models.input.NodeInput
+import edu.ie3.powerFactory2psdm.converter.NodeConverter
+import edu.ie3.powerFactory2psdm.model.Subnet
+import edu.ie3.powerFactory2psdm.model.powerfactory.PowerFactoryGrid.Nodes
 import edu.ie3.powerFactory2psdm.model.powerfactory.{
   PowerFactoryGrid,
   PowerFactoryGridMaps
 }
 import edu.ie3.powerFactory2psdm.util.GridPreparator
+
+import java.util.UUID
 
 /**
   * Functionalities to transform an exported and then parsed PowerFactory grid to the PSDM.
@@ -31,7 +37,22 @@ case object GridConverter {
     val graph = GridGraphBuilder.build(pfGridMaps)
     val subnets = SubnetBuilder.buildSubnets(graph, pfGridMaps.uuid2Node)
     val psdmNodes = subnets.flatMap(
-      subnet => NodeConverter.convertNodesOfSubnet(subnet, pfGridMaps.uuid2Node)
+      subnet => convertNodesOfSubnet(subnet, pfGridMaps.uuid2Node)
     )
+  }
+
+  /**
+    * Converts all nodes within a subnet to PSDM [[NodeInput]]
+    *
+    * @param subnet the subnet with reference to all PF nodes that live within
+    * @param uuid2node map that connects uuids with the associate PF [[Nodes]]
+    * @return list of all converted [[NodeInput]]
+    */
+  def convertNodesOfSubnet(
+      subnet: Subnet,
+      uuid2node: Map[UUID, Nodes]
+  ): List[NodeInput] = {
+    (for (nodeUUID <- subnet.nodeUuids)
+      yield NodeConverter.convertNode(nodeUUID, uuid2node, subnet)).toList
   }
 }
